// 左家辉月考/左家辉yk/zjhyk/src/type/type.ts
export interface detailface {
  ind: number;
  inds: number;
  navlist: Array<any>;
}
export interface dingdanface {
  visible: boolean;
}
export interface dizhiface {
  titles: string;
  names: string;
  phones: string;
  list: Array<any>;
  ids: any;
}
export interface homelistface {
  hea: Array<any>;
  list: any;
}